src/playground/pyproject.toml
[project]
name = "playground"
description = "Playground for running Chainlit apps."
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk",
    "chainlit",
    "httpx",
]
